Add test_discount_with_quantity()

+++ extensions/amortizable-sale-ext/src/api.rs
@@ -45,7 +45,7 @@ pub mod input {
     #[derive(Clone, Debug, Deserialize, PartialEq)]
     #[serde(rename_all(deserialize = "camelCase"))]
     pub struct Cost {
-        pub amount: Decimal,
+        pub amount: Int,
         pub currency_code : CurrencyCode,
     }
 
@@ -88,7 +88,7 @@ impl Default for DiscountConfiguration {
         DiscountConfiguration {
             discount_requirement_type: DiscountRequirementType::Subtotal,
             rules: vec![Rule {
-                value: RuleValue::FixedAmount { value: 10.00, amount_or_quantity: 50.00 },
+                value: RuleValue::FixedAmount { value: 10.00, amount_or_quantity: 50 },
             }],
         }
     }
@@ -204,13 +204,13 @@ pub enum RuleValue {
         #[serde_as(as = "DisplayFromStr")]
         value: Decimal,
         #[serde_as(as = "DisplayFromStr")]
-        amount_or_quantity: Decimal,
+        amount_or_quantity: Int,
     },
     Percentage {
         #[serde_as(as = "DisplayFromStr")]
         value: Decimal,
         #[serde_as(as = "DisplayFromStr")]
-        amount_or_quantity: Decimal,
+        amount_or_quantity: Int,
     },
 }
